requirements.txt
openai
yt-dlp
pydub
openai-whisper